ROOM OCCUPANCY MONITORING AND SECURITY SYSTEM using ARM7 LPC2148
DESIGN
Room occupancy monitoring systems are used to detect the movement
detection in a room by PIR (passive infrared sensors). PIR motion
detector used to control an outdoor, automatic light. A PIR-based motion
detector is used to sense movement of people, animals, or other objects.
They are commonly used in burglar alarms and automatically-activated
lighting systems.

If any movement is detected then microcontroller power ON necessary
devices as controlled power delivery or if there is no movement in the
room then it saves power. RTC is enabled for the controller as a result
the circuit is designed to detect the night time and after the day time,
the system automatically changes to a security system and in such a
condition when a movement is detected in the room, security alert is ON
and sends a alert message to a predetermined phone number.